добавил авторизацию и аутентификацию, реализовал ограничения для ролей

+++ src/main/java/ru/skypro/homework/service/AdService.java
@@ -6,8 +6,10 @@ import ru.skypro.homework.dto.AdDTO;
 import ru.skypro.homework.dto.AdsDTO;
 import ru.skypro.homework.dto.CreateOrUpdateAdDTO;
 import ru.skypro.homework.dto.ExtendedAdDTO;
+import ru.skypro.homework.exception.NoAccessException;
 import ru.skypro.homework.exception.UserNotAuthorizedException;
 import ru.skypro.homework.model.Ad;
+import ru.skypro.homework.model.Role;
 import ru.skypro.homework.model.User;
 import ru.skypro.homework.service.dao.AdDAO;
 import ru.skypro.homework.service.mapper.AdMapper;
@@ -37,7 +39,7 @@ public class AdService {
         newAd.setTitle(properties.getTitle());
         newAd.setPrice(properties.getPrice());
         newAd.setAuthor(user);
-        newAd.setImage(imageService.uploadAdImage(image));
+        newAd.setImage(imageService.uploadAdImage(properties.getTitle(), image));
 
         adDAO.addAd(newAd);
 
@@ -65,19 +67,37 @@ public class AdService {
         return adMapper.adToExtendedAdDTO(adDAO.getAdById(id));
     }
 
-    public void deleteAd(int id) {
+    public void deleteAd(int id) throws UserNotAuthorizedException, NoAccessException {
+        User user = userService.getAuthorizedUser();
         Ad ad = adDAO.getAdById(id);
-        adDAO.removeAd(ad);
+
+        if (ad.getAuthor().getId() == user.getId() || user.getRole().equals(Role.ADMIN)) {
+            adDAO.removeAd(ad);
+        } else {
+            throw new NoAccessException("No access");
+        }
     }
 
-    public AdDTO updateAd(int id, CreateOrUpdateAdDTO adDTO) {
+    public AdDTO updateAd(int id, CreateOrUpdateAdDTO adDTO) throws UserNotAuthorizedException, NoAccessException {
+        User user = userService.getAuthorizedUser();
         Ad ad = adDAO.getAdById(id);
-        ad.setDescription(adDTO.getDescription());
-        ad.setPrice(adDTO.getPrice());
-        ad.setTitle(ad.getTitle());
 
+        if (ad.getAuthor().getId() == user.getId() || user.getRole().equals(Role.ADMIN)) {
+            ad.setDescription(adDTO.getDescription());
+            ad.setPrice(adDTO.getPrice());
+            ad.setTitle(ad.getTitle());
+            adDAO.updateAd(ad);
+            return adMapper.adToAdDTO(ad);
+        } else {
+            throw new NoAccessException("No access");
+        }
+
+    }
+
+    public byte[] updateAdImage(int adId, MultipartFile image) throws UserNotAuthorizedException, IOException {
+        Ad ad = adDAO.getAdById(adId);
+        ad.setImage(imageService.uploadAdImage(ad.getTitle(), image));
         adDAO.updateAd(ad);
-
-        return adMapper.adToAdDTO(ad);
+        return image.getBytes();
     }
 }
